Add scripts for working with node.js.
Create scripts for invoking the JavaScript file generated by
ClojureScript in the node.js environment. In addition, create a script
for creating a node.js REPL in that same environment.

This scripts are combinations of the files used to build the application
and to create a REPL and the scripts available from the ClojureScript
Quick Start (../cljs_qs) project.