Began trying to read the file contents using node.
The code is not working. I'm uncertain if this is because of the changes
I've made to time-calc/core or if I have an environment error.

I am checking in this code. I then will return to a previous commit to
test the environment. (I should have done this earlier.)